+++ tests/testthat/test-find_layer.R
@@ -1,6 +1,6 @@
 source(file = "../load_testdata.R")
 
-index <- get_uid_list(test_docs)
+index <- make_index(test_docs)
 proj_id <- index$identifier[index$category == "Project"]
 layer_categories = c("Feature", names(config$categories$Feature$trees))