+++ tests/testthat/test-gather_trenches.R
@@ -1,6 +1,6 @@
 source(file = "../load_testdata.R")
 
-uidlist <- get_uid_list(test_docs)
+uidlist <- make_index(test_docs)
 
 test_that("returns same amount of elements", {
   expect_equal(length(gather_trenches(uidlist)), nrow(uidlist))